Split regen logic into another script.
